finally got the twig extension to register, but now I'm receiving two error messages:   - "Fatal error: Call to undefined function twig_trim_string_filter() in /drupal-8.0.0-beta9/core/lib/Drupal/Core/Template/TwigEnvironment.php(120) : eval()'d code on line 40" which doesn't make sense to me because the watchdog error is stating the following error.   - "Warning: explode() expects parameter 2 to be string, array given in twig_trim_string_filter() (line 306 of /Users/erik.baldwin/Sites/DAMP/drupal-8.0.0-beta9/core/themes/engines/twig/twig.engine"
I'm wondering what I'm doing wrong here. I had to move the function that is called by the twig filter outside of the class declaration in order for it to be registered, and I'm not sure why that has to be. I'm also not certain to why the first arg passed into the function isn't available, or an actual string, like it should be.